Put log file in .conf
I can now put the log file in the .conf file, so I can test a client
without polluting the production backup log file.

Also, I added the client executables and the test files to .gitignore.
For the test files, I can put them into git in the future if I want to,
but for now I expect if I were writing test scripts, I'd create the test
files on the fly.