Using new approach. (Not working yet.)
Changed ACLs to a config option, because they're not supported on the DSM.
Added to .gitignore.
Need to create the folder for the backup client before I run the backup.
Maybe this can be done in here if needed, but we start to need to address
the permissions issue. Right now the login is with admin and needs a password.