Another partial working on restore.
Restore requires sudo if the files being retrieved aren't owned by the
user running restore. I may also have to figure out --super and --fake-super.
I really don't think I've thought through the syntax of restoring.

But it does work in a fashion. It gets the files back.

Backing up a path that starts with "./" is bizarre, and will likely cause
error messages on restore, because it will try to set date and/or owner
of the target directory.