Moved code and improved logic on full versus incremental backup.
A full backup just means not passing in the link flags, so I implemented it.
I moved all code into functions, so it's a little more organized. It only
does timing on backups, not restore or ls. (It might make sense to report
time for restore as well.)

My earlier marvelling that it was working without su privileges was
misplaced. It's not doing the right thing. Need to work on that.