Something that works for me.
This commit was deployed to the minecraft server, and is now running. Need
to watch it run for a few days.

Assumes a default configuration for the Synology. Would be insecure in
anything other than a home network.

Requires manual set-up on the Synology side to make the directory, and
create and populate the password file.

Discovered that the "ls" format from rsync changes from version to version.