Add verbose flag.
Added verbose flag, and re-implemented the debugging output.

This was to fix why it wasn't working in real life. It turns out that
when I back up from /, the date on the backup directories comes from
the mod time of /, which is constant. So I wasn't getting the right
directory as the newest, and so I was making a full backup on every
backup.